Update docusaurus to 2.0.0-beta.15
A lot of changes since `-beta.0`. I'm pulling this in to resolve the
dependabot auto-PRs en masse.

Updated our TargetVersionFilled darkmode sensitive function with the API
changes in docusaurus.

Update `.nvmrc` to a version compatible with this version of docusaurus
(happens to be my local one, `node -v > .nvmrc`).

Also update the .forestry image to a node v14 one (I don't know what
that's used for but 🤷)