Add a small script for checking page ids
This script checks if a page `id: ` tag matches the basename (without
extension) of the file, and returns non-zero if there's a mismatch.

This helps reduce confusion IMO, and I didn't see a reason to not
require it.

Add the check to github actions.

Correcting the ones that don't match by following this strategy:

- rename the _file_ (NOT the ID) to match the ID. this preserves
  external links, since the slug is based on the ID. the file name is
  not used in the generated site.
- update any cross links here that use the file name to match the new
  name (docusaurus fails to compile when they're broken 😍)
- run prettier (npx prettier --write $(gd HEAD --name-only))