Update docs for hawkBit API
 ### Issues addressed
The previous instructions relied on the General HTTP API. We want the
instructions to use our hawkBit API.

 ### Summary of changes
Add settings for hawkBit.

I grabbed the `swupdate.cfg` settings that we're using for our Yocto
example; it's unclear which of these are actually required. Once we have
a CI test in place (working on it!), it'll be easier to modify these
variables and determine which are required and which are not.

FUTURE WORK: Just after this section is "Integration with Yocto". That
has not yet been pointed at our new sdk docs; the docs will need some
cleanup before this should be performed (specifically, the readme (or a
new readme) should be updated to be more from the "implementer"
point-of-view).

 ### Test plan
Using the current yocto implementation as the test guinea pig, which
roughly follows these instructions.